+++ src/app/overlays/dialogs/api-request/api-request-dialog.component.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+import { MatDialogRef } from '@angular/material/dialog';
+
+@Component({
+    selector: 'app-api-request-dialog',
+    template: ` <app-api-request-overlay (close)="closeDialog()"></app-api-request-overlay> `,
+})
+export class ApiRequestDialogComponent {
+    constructor(private readonly _dialogRef: MatDialogRef<ApiRequestDialogComponent>) {}
+
+    closeDialog(): void {
+        this._dialogRef.close();
+    }
+}