+++ src/app/pages/dashboard/dashboard.component.ts
@@ -1,6 +1,6 @@
 import { Component, ViewEncapsulation } from '@angular/core';
 import * as Colors from '@brightlayer-ui/colors';
-import { Router } from '@angular/router';
+import { ActivatedRoute, Params, Router } from '@angular/router';
 import { UtilService } from '@app/services/util.service';
 import { MatDialog } from '@angular/material/dialog';
 import { ViewportService } from '@app/services/viewport.service';
@@ -26,6 +26,8 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 import { ReceiveBottomSheetComponent } from '@app/overlays/bottom-sheet/receive/receive-bottom-sheet.component';
 import { ReceiveDialogComponent } from '@app/overlays/dialogs/receive/receive-dialog.component';
 import { ReceiveOverlayData } from '@app/overlays/actions/receive/receive.component';
+import { ApiRequestBottomSheetComponent } from '@app/overlays/bottom-sheet/api-request/api-request-bottom-sheet.component';
+import { ApiRequestDialogComponent } from '@app/overlays/dialogs/api-request/api-request-dialog.component';
 import { animate, style, transition, trigger } from '@angular/animations';
 
 @UntilDestroy()
@@ -53,6 +55,7 @@ export class DashboardComponent {
 
     constructor(
         public vp: ViewportService,
+        private readonly _route: ActivatedRoute,
         private readonly _router: Router,
         private readonly _dialog: MatDialog,
         private readonly _util: UtilService,
@@ -68,6 +71,25 @@ export class DashboardComponent {
                 this.totalBalance = this._util.numberWithCommas(store.totalBalance);
             }
         });
+        this._route.queryParams.subscribe((params) => this._checkForApiRequestViaQueryParams(params));
+    }
+
+    private _checkForApiRequestViaQueryParams(params: Params): void {
+        if (!this._isValidParams(params)) {
+            if (params) {
+                console.warn('Invalid query parameters provided while creating a BananoStand API request.');
+                console.warn('Parameter Options: "request (send | change), address, amount (send only)');
+                console.warn('Parameters provided: ', params);
+            }
+            return;
+        }
+        if (this.vp.sm) {
+            setTimeout(() => {
+                this._sheet.open(ApiRequestBottomSheetComponent);
+            }, this.bottomSheetOpenDelayMs);
+        } else {
+            this._dialog.open(ApiRequestDialogComponent);
+        }
     }
 
     openEnterSeedOverlay(): void {
@@ -201,5 +223,16 @@ export class DashboardComponent {
     toggleDashboardDisplay(): void {
         CHANGE_PREFERRED_DASHBOARD_VIEW.next(this.showTableView ? 'card' : 'table');
         this.accountActionsOverlayOpen = false;
+    }
+
+    private _isValidParams(params: Params): boolean {
+        if (!params || !params.address || !params.request) {
+            return false;
+        }
+        const request = params.request.toLowerCase();
+        if (request === 'send') {
+            return !isNaN(Number(params.amount)) && this._util.isValidAddress(params.address);
+        }
+        return this._util.isValidAddress(params.address);
     }
 }